Feature/record frag

+++ app/src/main/java/com/hapataka/questwalk/domain/entity/HistoryEntity.kt
@@ -8,11 +8,11 @@ sealed class HistoryEntity(open val registerAt: String = "") {
         val quest: String = "",
         val time: Long = 0,
         val distance: Float = 0f,
-        val step: Int = 0,
+        val step: Long = 0,
         @JvmField
         val isFailed: Boolean = false,
         val locations: List<Pair<Float,Float>>?,
-        val questLocation: Pair<Float, Float>? = null,
+        val questLocation: Pair<Float,Float>? = null,
         val questImg: String? = null,
         val type: String = RESULT_TYPE
     ) : HistoryEntity(registerAt)
